Correction ejercicio004
